+++ dlrover/python/tests/test_diagnosis.py
@@ -18,7 +18,6 @@ from dlrover.python.common.constants import NodeStatus
 from dlrover.python.diagnosis.common.constants import (
     DiagnosisActionType,
     DiagnosisConstant,
-    DiagnosisDataType,
 )
 from dlrover.python.diagnosis.common.diagnosis_action import (
     DiagnosisAction,
@@ -26,8 +25,6 @@ from dlrover.python.diagnosis.common.diagnosis_action import (
     EventAction,
     NodeAction,
 )
-from dlrover.python.diagnosis.common.diagnosis_data import TrainingLog
-from dlrover.python.master.diagnosis.diagnosis import DiagnosisDataManager
 
 
 class DiagnosisTest(unittest.TestCase):
@@ -36,23 +33,6 @@ class DiagnosisTest(unittest.TestCase):
 
     def tearDown(self):
         pass
-
-    def test_data_manager(self):
-        mgr = DiagnosisDataManager(5)
-        log1 = TrainingLog(0)
-        mgr.store_data(log1)
-        time.sleep(1)
-        log2 = TrainingLog(0)
-        mgr.store_data(log2)
-
-        logs = mgr.get_data(DiagnosisDataType.TRAINING_LOG)
-        self.assertEqual(len(logs), 2)
-
-        time.sleep(6)
-        log3 = TrainingLog(0)
-        mgr.store_data(log3)
-        logs = mgr.get_data(DiagnosisDataType.TRAINING_LOG)
-        self.assertEqual(len(logs), 1)
 
     def test_action_basic(self):
         basic_action = DiagnosisAction()
