+++ dlrover/python/tests/test_diagnosis_manager.py
@@ -1,0 +1,54 @@
+# Copyright 2024 The DLRover Authors. All rights reserved.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import time
+import unittest
+
+from dlrover.python.diagnosis.common.constants import DiagnosisDataType
+from dlrover.python.diagnosis.common.diagnosis_data import TrainingLog
+from dlrover.python.master.diagnosis.diagnosis_data_manager import (
+    DiagnosisDataManager,
+)
+from dlrover.python.master.diagnosis.diagnosis_manager import DiagnosisManager
+
+
+class DiagnosisTest(unittest.TestCase):
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    def test_data_manager(self):
+        mgr = DiagnosisDataManager(5)
+        log1 = TrainingLog(0)
+        mgr.store_data(log1)
+        time.sleep(1)
+        log2 = TrainingLog(0)
+        mgr.store_data(log2)
+
+        logs = mgr.get_data(DiagnosisDataType.TRAINING_LOG)
+        self.assertEqual(len(logs), 2)
+
+        time.sleep(6)
+        log3 = TrainingLog(0)
+        mgr.store_data(log3)
+        logs = mgr.get_data(DiagnosisDataType.TRAINING_LOG)
+        self.assertEqual(len(logs), 1)
+
+    def test_diagnosis_manager(self):
+        mgr = DiagnosisManager()
+
+        mgr.pre_check()
+        mgr.start_observing()
+        mgr.stop_observing()
