+++ dlrover/python/util/time_util.py
@@ -15,6 +15,11 @@ from datetime import datetime, timedelta
 
 
 def has_expired(timestamp: float, time_period: float) -> bool:
+    """
+    :param timestamp: data timestamp
+    :param time_period: expire time period in seconds
+    :return: indicate if the timestamp has expired
+    """
     dt = datetime.fromtimestamp(timestamp)
-    expired_dt = dt + timedelta(milliseconds=time_period)
+    expired_dt = dt + timedelta(milliseconds=time_period * 1000)
     return expired_dt < datetime.now()
